requirements.txt
fastapi
pydantic
xgboost
fastapi
uvicorn
pandas
numpy
joblib
wandb
python-dotenv
scikit-learn